Internalize automining, add new mining endpoints, optimize shutdowns
Automining is now internal within the backend, with a built-in toggle. Instead of relying on the frontend to control mining operations entirely.
One mining endpoint has been depreciated, and two new mining endpoints have been added, alongside two new unrelated endpoints: Ping and Stop. For pinging the server to check that it's alive, and stopping the server (The frontend makes use of the Stop endpoint accordingly for unified shutdowns).