Recursively sync missing blocks
Previously, nodes only had a "init" phase in which they'd connect to all peers and request for blocks + transactions simultaniously. The problem with this, is that consumer-grade devices were unable to continously stay synced, as they can only communicate outwards, not accept connections from the outside.

This allows those nodes to continously stay synced with it's peers, with or without a portforwarded configuration.